requirements.txt
Django==1.9.8
django-filter==1.0.0
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
psycopg2==2.6.2